+++ libft.h
@@ -12,7 +12,7 @@ void                *ft_memccpy(void *dest, const void *src, int c, size_t n);
 void                *ft_memmove(void *dest, const void *src, size_t len);
 void                *ft_memchr(const void *s, int c, size_t n);
 int                 ft_memcmp(const void *s1, const void *s2, size_t n);
-int                 ft_strlen(char const *str);
+size_t              ft_strlen(char const *str);
 int                 ft_isalpha(int c);
 int                 ft_isdigit(int c);
 int                 ft_isalnum(int c);
